+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "primeflex": "^2.0.0",
+    "primeicons": "^4.1.0",
     "primereact": "^6.0.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
