+++ src/App.css
@@ -1,7 +1,7 @@
 
 /* LANDING PAGE */
 .landing-page {
-  background-color: #E7E5E5;;
+  background-color: #FFFFFF;;
   color: black;
   display: grid;
   height: 100vh;
@@ -20,14 +20,33 @@
   grid-row: 3 / span 2;
 }
 
+/* AUTH CONTAINER */
+.auth-container {
+  background-color: #FFFFFF;
+  grid-column: 7 / -2;
+  grid-row: 2 / -2;
+  display: grid;
+  grid-gap: 1%;
+  grid-template-columns: repeat(12, 1fr);
+  grid-template-rows: repeat(12, 1fr);
+}
+
+.log-in-button {
+  grid-column: 4 / span 2;
+  grid-row: 2 / span 1;
+}
+
+.sign-up-button {
+  grid-column: 8 / span 2;
+  grid-row: 2 / span 1;
+}
+
 .landing-nav {
-  grid-column: 4 / span 6;
+  grid-column: 2 / span 2;
   grid-row: 5 / 6;
   display: flex;
   justify-content: space-evenly;
 }
-
-/* LOGIN PAGE */
