+++ src/landing/LogIn.js
@@ -1,0 +1,11 @@
+import React, { useState } from 'react'
+
+const LogIn = () => {
+    return (
+      <>
+        <h1>Sign In</h1>
+      </>
+    );
+}
+
+export default LogIn
