--- src/login/LogInContent.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const LogInContent = () => {
-    return (
-        <h1>log in content</h1>
-    )
-}
-
-export default LogInContent
