--- src/login/LogInInputs.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-const LogInInputs = () => {
-    return (
-      <>
-        <form action="login">
-          <label htmlFor="login">username</label>
-          <input type="text" name="" id="" />
-        </form>
-        <a href="/">Back</a>
-      </>
-    );
-}
-
-export default LogInInputs
