--- src/login/LogInPage.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import LogInContent from './LogInContent'
-import LogInInputs from './LogInInputs'
-
-const LoginPage = () => {
-    return (
-      <>
-        <LogInContent />
-        <LogInInputs />
-      </>
-    );
-}
-
-export default LoginPage
